+++ src/test/java/net/binis/codegen/AnnotationsTest.java
@@ -23,29 +23,29 @@ package net.binis.codegen;
 import lombok.extern.slf4j.Slf4j;
 import net.binis.codegen.generation.core.Helpers;
 import net.binis.codegen.test.BaseTest;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 @Slf4j
-public class AnnotationsTest extends BaseTest {
+class AnnotationsTest extends BaseTest {
 
-    @Before
+    @BeforeEach
     public void cleanUp() {
         Helpers.cleanUp();
     }
 
     @Test
-    public void test() {
+    void test() {
         testSingle("annotation/default1.java", "annotation/default1-0.java", "annotation/default1-1.java");
     }
 
     @Test
-    public void testFor() {
+    void testFor() {
         testSingle("annotation/testFor1.java", "annotation/testFor1-0.java", "annotation/testFor1-1.java");
     }
 
     @Test
-    public void testForDefaultMethod() {
+    void testForDefaultMethod() {
         testSingle("annotation/testForDefault1.java", null, "annotation/testForDefault1-1.java");
     }