+++ src/test/java/net/binis/codegen/BaseModifierTest.java
@@ -23,24 +23,24 @@ package net.binis.codegen;
 import lombok.extern.slf4j.Slf4j;
 import net.binis.codegen.generation.core.Helpers;
 import net.binis.codegen.test.BaseTest;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 @Slf4j
-public class BaseModifierTest extends BaseTest {
+class BaseModifierTest extends BaseTest {
 
-    @Before
+    @BeforeEach
     public void cleanUp() {
         Helpers.cleanUp();
     }
 
     @Test
-    public void test() {
+    void test() {
         testSingle("base/modifier/baseTest1.java", "base/modifier/baseTest1-0.java", "base/modifier/baseTest1-1.java");
     }
 
     @Test
-    public void testAsync() {
+    void testAsync() {
         testSingle("base/modifier/baseTest2.java", "base/modifier/baseTest2-0.java", "base/modifier/baseTest2-1.java");
     }