+++ src/test/java/net/binis/codegen/BasicsTest.java
@@ -24,31 +24,31 @@ import lombok.extern.slf4j.Slf4j;
 import net.binis.codegen.generation.core.Helpers;
 import net.binis.codegen.test.BaseTest;
 import org.apache.commons.lang3.tuple.Triple;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
 @Slf4j
-public class BasicsTest extends BaseTest {
+class BasicsTest extends BaseTest {
 
-    @Before
+    @BeforeEach
     public void cleanUp() {
         Helpers.cleanUp();
     }
 
     @Test
-    public void test() {
+    void test() {
         testSingle("basic/Test1.java", "basic/Test1-0.java", "basic/Test1-1.java");
     }
 
     @Test
-    public void testForAnnotations() {
+    void testForAnnotations() {
         testSingle("basic/TestFor.java", "basic/TestFor-0.java", "basic/TestFor-1.java");
     }
 
     @Test
-    public void testExtended() {
+    void testExtended() {
         testMulti(List.of(
                 Triple.of("extended/Extended1.java", "extended/Extended1-0.java", "extended/Extended1-1.java"),
                 Triple.of("extended/Test1.java", "extended/Test1-0.java", "extended/Test1-1.java")