+++ src/test/java/net/binis/codegen/prototype/CompiledPrototype.java
@@ -25,7 +25,7 @@ import net.binis.codegen.annotation.CodePrototype;
 import net.binis.codegen.annotation.Default;
 import net.binis.codegen.spring.annotation.QueryFragment;
 
-import javax.persistence.OneToOne;
+import jakarta.persistence.OneToOne;
 import java.util.List;
 
 @CodePrototype(generateImplementation = false)