+++ src/test/resources/generics/Test5-2.java
@@ -1,10 +1,10 @@
 /*Generated code by Binis' code generator.*/
 package net.binis.codegen;
 
-import net.binis.codegen.factory.CodeFactory;
 import net.binis.codegen.test.TestExecutor;
 import net.binis.test.card.AccountOverviewCard;
-import static org.junit.Assert.*;
+
+import static org.junit.jupiter.api.Assertions.*;
 
 public class Execute extends TestExecutor {